// rrs-client/src/app/shared/enums/FormType.ts
/**
 * Az űrlap típus felsorlási típus
 */
export enum FormType {
    EDIT_EVENT_RESERVATION_FORM = "EDIT_EVENT_RESERVATION_FORM",
    EDIT_CLASS_RESERVATION_FORM = "EDIT_CLASS_RESERVATION_FORM",
    OBSERVE_EVENT_RESERVATION_FORM = "OBSERVE_EVENT_RESERVATION_FORM",
    OBSERVE_CLASS_RESERVATION_FORM = "OBSERVE_CLASS_RESERVATION_FORM",
    EDIT_BUILDING_FORM = "EDIT_BUILDING_FORM",
    EDIT_CLASSROOM_FORM = "EDIT_CLASSROOM_FORM",
    EDIT_SEMESTER_FORM = "EDIT_SEMESTER_FORM",
    EDIT_SUBJECT_FORM = "EDIT_SUBJECT_FORM",
    EDIT_USER_FORM = "EDIT_USER_FORM",
}